Add implementation for data preservation
