+++ components/nameable.rb
