+++ components/student.rb
@@ -1,4 +1,4 @@
-require './person'
+require_relative 'person'
 
 class Student < Person
   attr_reader :classroom
@@ -14,5 +14,10 @@ class Student < Person
 
   def play_hooky
     '¯\(ツ)/¯'
+  end
+
+  def to_json(*_args)
+    { 'id' => @id, 'name' => @name, 'age' => @age, 'parent_permission' => @parent_permission,
+      'class' => self.class.name, 'rentals' => @rentals }
   end
 end
