+++ components/teacher.rb
@@ -1,4 +1,4 @@
-require './person'
+require_relative 'person'
 
 class Teacher < Person
   def initialize(age, name, specialization, parent_permission: true, person_type: 'teacher')
@@ -8,5 +8,10 @@ class Teacher < Person
 
   def can_use_services?
     true
+  end
+
+  def to_json(*_args)
+    { 'id' => @id, 'name' => @name, 'age' => @age, 'specialization' => @specialization, 'class' => self.class.name,
+      'rentals' => @rentals }
   end
 end
