+++ data/_books.json
@@ -1,0 +1,14 @@
+[
+    {
+      "title": "The Great Gatsby",
+      "author": "F. Scott Fitzgerald"
+    },
+    {
+      "title": "To Kill a Mockingbird",
+      "author": "Harper Lee"
+    },
+    {
+      "title": "1984",
+      "author": "George Orwell"
+    }
+  ]
