+++ data/books.json
@@ -1,0 +1,9 @@
+[
+  {
+    "title": "The great gatsby",
+    "author": "F. scott fitzgerald",
+    "rentals": [
+
+    ]
+  }
+]
