+++ data/rentals.json
@@ -1,0 +1,9 @@
+[
+  {
+    "date": "2023-02-08",
+    "title": "The great gatsby",
+    "name": "Mancoba",
+    "id": 195,
+    "class": "Student"
+  }
+]
